requirements.txt
crewai
python-dotenv
requests
openai
langchain
langchain-openai
